+++ CHANGELOG.md
@@ -4,7 +4,18 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.9.14 - Unreleased
+## Version 0.9.15 - Unreleased
+
+### Added
+* `center_x` and `center_y` to Box and Boxes.
+* `Boxes.resize` now has `about` argument.
+
+### Fixed
+* Bug in `warp_affine` when the input image has more than 4 dims, has an empty dimension, and `border_value` is an array.
+* Issue in `nodata_checkerboard` with masked arrays.
+
+
+## Version 0.9.14 - Released 2023-03-16
 
 ### Added
 * New function: `exactly_1channel`.